refactor: move Vacancy & VacancyRecord to shared code;
