+++ internal/server/core/model.go
@@ -1,5 +1,7 @@
 package core
 
+import shcore "github.com/ghazlabs/idn-remote-entry/internal/shared/core"
+
 type SubmitType string
 
 const (
@@ -9,25 +11,10 @@ const (
 
 type SubmitRequest struct {
 	SubmissionType SubmitType `json:"submission_type"`
-	Vacancy
+	shcore.Vacancy
 }
 
 func (r SubmitRequest) Validate() error {
 	// TODO
 	return nil
-}
-
-type Vacancy struct {
-	JobTitle         string   `json:"job_title"`
-	CompanyName      string   `json:"company_name"`
-	CompanyLocation  string   `json:"company_location"`
-	ShortDescription string   `json:"short_description"`
-	RelevantTags     []string `json:"relevant_tags"`
-	ApplyURL         string   `json:"apply_url"`
-}
-
-type VacancyRecord struct {
-	ID string `json:"id"`
-	Vacancy
-	PublicURL string `json:"public_url"`
 }
